The nrf_wifi source code moved to mdepx/lib/nrf_wifi as the nrfxlib does not host it anymore. Update nrfxlib.

+++ src/main.c
@@ -312,7 +312,7 @@ static int
 lte_wait(void)
 {
 	uint32_t cell_id;
-	uint16_t status;
+	int status;
 	int err;
 
 	printf("Awaiting registration in the LTE-M network...\n");
